requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
future==0.18.2
numpy==1.18.1
Pillow==7.0.0
pycparser==2.19
python-telegram-bot==12.4.1
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
